requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
streamlit
plotly
pandas
requests
httpx
python-multipart
